feat(types): replace 591 std:: type uses with datapod:: types
- Replace std::size_t → datapod::usize (456 uses)
- Replace std::ptrdiff_t → datapod::isize (42 uses)
- Replace std::uint8_t → datapod::u8 (39 uses)
- Replace std::uint32_t → datapod::u32 (35 uses)
- Replace std::uint64_t → datapod::u64 (11 uses)
- Replace std::uint16_t → datapod::u16 (5 uses)
- Replace std::int16_t → datapod::i16 (3 uses)
- Replace std::int64_t → datapod::i64 (3 uses)

Key changes:
- Add types/types.hpp include to 60+ files
- Fix datapod::usize to match std::size_t (unsigned long on 64-bit)
- Fix std::index_sequence parameter packs to use std::size_t
- Remove conflicting 'namespace dp = datapod' aliases
- Fix type mismatches in serialization and tuple code

All 118 tests passing!

+++ include/datapod/types/primitives.hpp
@@ -25,14 +25,16 @@ namespace datapod {
     using u64 = unsigned long long; // Always 64 bits
 
 // Size types - platform dependent
+// Note: We use the same underlying types as std::size_t/std::ptrdiff_t
+// to ensure compatibility with standard library functions
 #if defined(__LP64__) || defined(_WIN64)
-                                    // 64-bit platforms
-    using usize = unsigned long long;
-    using isize = signed long long;
+    // 64-bit platforms
+    using usize = unsigned long; // Same as std::size_t on 64-bit
+    using isize = signed long;   // Same as std::ptrdiff_t on 64-bit
 #else
-                                    // 32-bit platforms
-    using usize = unsigned int;
-    using isize = signed int;
+    // 32-bit platforms
+    using usize = unsigned int; // Same as std::size_t on 32-bit
+    using isize = signed int;   // Same as std::ptrdiff_t on 32-bit
 #endif
 
     // ============================================================================
